// frontend/assignments/3/src/Types/SummarizerType.ts
// Generated by https://quicktype.io

export interface SummarizedStock {
  id: number;
  maxProfit: number;
  buyIndex: number;
  sellIndex: number;
  company: string;
  buyDate: string;
  sellDate: string;
  buyPrice: number;
  sellPrice: number;
}

export interface SummarizedStockContent {
  maxProfit: number;
  buyIndex: number;
  sellIndex: number;
  company: string;
  buyDate: string;
  sellDate: string;
  buyPrice: number;
  sellPrice: number;
}

export interface SummarizedStockSliceType {
  summarizedStocks: SummarizedStock[];
  loading: boolean;
  error: string;
}

export interface StockData {
  date: string;
  prices: number[];
}

export interface ParsedStockData {
  company: string;
  data: {
    prices: number[];
    dayToIndexMap: {
      date: string;
      indexMap: number[];
    }[];
  };
}

export interface workerResult {
  company: string;
  buyDate: string;
  sellDate: string;
  buyPrice: number;
  sellPrice: number;
  maxProfit: number;
  buyIndex: number;
  sellIndex: number;
}